Refactored NodeEditorWindow organization structure, adjusted module import path and added an empty __init__.py file
